CELEBA-CON INSIGHTFACE/DATA_APIX/requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.24.0
insightface>=0.7.3
scikit-image>=0.21.0
pillow>=10.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.31.0
gdown>=4.7.1
scipy>=1.11.0